+++ components/net/sf/briar/db/SynchronizedDatabaseComponent.java
@@ -11,6 +11,7 @@ import java.util.logging.Logger;
 
 import net.sf.briar.api.ContactId;
 import net.sf.briar.api.Rating;
+import net.sf.briar.api.db.DatabaseListener;
 import net.sf.briar.api.db.DbException;
 import net.sf.briar.api.db.NoSuchContactException;
 import net.sf.briar.api.protocol.Ack;
@@ -146,7 +147,7 @@ class SynchronizedDatabaseComponent<Txn> extends DatabaseComponentImpl<Txn> {
 			}
 		}
 		// Call the listeners outside the lock
-		if(added) callMessageListeners();
+		if(added) callListeners(DatabaseListener.Event.MESSAGES_ADDED);
 	}
 
 	public void findLostBatches(ContactId c) throws DbException {
@@ -550,7 +551,7 @@ class SynchronizedDatabaseComponent<Txn> extends DatabaseComponentImpl<Txn> {
 			}
 		}
 		// Call the listeners outside the lock
-		if(anyAdded) callMessageListeners();
+		if(anyAdded) callListeners(DatabaseListener.Event.MESSAGES_ADDED);
 	}
 
 	public void receiveOffer(ContactId c, Offer o, RequestWriter r)
@@ -691,27 +692,37 @@ class SynchronizedDatabaseComponent<Txn> extends DatabaseComponentImpl<Txn> {
 
 	public void subscribe(Group g) throws DbException {
 		if(LOG.isLoggable(Level.FINE)) LOG.fine("Subscribing to " + g);
+		boolean added = false;
 		synchronized(subscriptionLock) {
 			Txn txn = db.startTransaction();
 			try {
-				db.addSubscription(txn, g);
+				if(!db.containsSubscription(txn, g.getId())) {
+					db.addSubscription(txn, g);
+					added = true;
+				}
 				db.commitTransaction(txn);
 			} catch(DbException e) {
 				db.abortTransaction(txn);
 				throw e;
 			}
 		}
+		// Call the listeners outside the lock
+		if(added) callListeners(DatabaseListener.Event.SUBSCRIPTIONS_UPDATED);
 	}
 
 	public void unsubscribe(GroupId g) throws DbException {
 		if(LOG.isLoggable(Level.FINE)) LOG.fine("Unsubscribing from " + g);
+		boolean removed = false;
 		synchronized(contactLock) {
 			synchronized(messageLock) {
 				synchronized(messageStatusLock) {
 					synchronized(subscriptionLock) {
 						Txn txn = db.startTransaction();
 						try {
-							db.removeSubscription(txn, g);
+							if(db.containsSubscription(txn, g)) {
+								db.removeSubscription(txn, g);
+								removed = true;
+							}
 							db.commitTransaction(txn);
 						} catch(DbException e) {
 							db.abortTransaction(txn);
@@ -721,5 +732,7 @@ class SynchronizedDatabaseComponent<Txn> extends DatabaseComponentImpl<Txn> {
 				}
 			}
 		}
+		// Call the listeners outside the lock
+		if(removed) callListeners(DatabaseListener.Event.SUBSCRIPTIONS_UPDATED);
 	}
 }
